refactor validate Submit with new ArgoFileValidator classes
