+++ file_checker_exec/src/main/java/fr/coriolis/checker/output/ResultsFile.java
@@ -25,10 +25,9 @@ import javax.xml.transform.stream.StreamSource;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-import fr.coriolis.checker.filetypes.ArgoDataFile;
-import fr.coriolis.checker.filetypes.ArgoDataFile.FileType;
-import fr.coriolis.checker.filetypes.ArgoProfileFileValidator;
-import fr.coriolis.checker.filetypes.ValidationResult;
+import fr.coriolis.checker.core.ArgoDataFile;
+import fr.coriolis.checker.core.ValidationResult;
+import fr.coriolis.checker.core.ArgoDataFile.FileType;
 import fr.coriolis.checker.specs.ArgoDate;
 
 public class ResultsFile {
@@ -457,7 +456,7 @@ public class ResultsFile {
 			metaStationParameters(argo);
 
 			if (formatPassed && argo.fileType() == FileType.PROFILE) {
-				addMetaPsalStats((ArgoProfileFileValidator) argo);
+				addMetaPsalStats(argo);
 			}
 
 		} else if (argo.fileType() == FileType.TRAJECTORY || argo.fileType() == FileType.BIO_TRAJECTORY) {
@@ -564,7 +563,7 @@ public class ResultsFile {
 		}
 	} // ..end metaData
 
-	public void addMetaPsalStats(ArgoProfileFileValidator argo) throws XMLStreamException {
+	public void addMetaPsalStats(ArgoDataFile argo) throws XMLStreamException {
 		// ...............report PSAL adjustment statistics...............
 		// ..assumes:
 		// ..- Argo-open was successful
