+++ file_checker_exec/src/main/java/fr/coriolis/checker/validators/ArgoFileValidator.java
@@ -1,4 +1,4 @@
-package fr.coriolis.checker.filetypes;
+package fr.coriolis.checker.validators;
 
 import java.io.IOException;
 import java.io.PrintStream;
@@ -12,7 +12,9 @@ import java.util.regex.Pattern;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-import fr.coriolis.checker.filetypes.ArgoDataFile.FileType;
+import fr.coriolis.checker.core.ArgoDataFile;
+import fr.coriolis.checker.core.ValidationResult;
+import fr.coriolis.checker.core.ArgoDataFile.FileType;
 import fr.coriolis.checker.specs.ArgoAttribute;
 import fr.coriolis.checker.specs.ArgoDate;
 import fr.coriolis.checker.specs.ArgoDimension;
