+++ file_checker_exec/src/main/java/fr/coriolis/checker/validators/ArgoMetadataFileValidator.java
@@ -1,4 +1,4 @@
-package fr.coriolis.checker.filetypes;
+package fr.coriolis.checker.validators;
 
 import java.io.IOException;
 import java.util.Date;
@@ -11,6 +11,7 @@ import java.util.regex.Pattern;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
+import fr.coriolis.checker.core.ArgoDataFile;
 import fr.coriolis.checker.specs.ArgoConfigTechParam;
 import fr.coriolis.checker.specs.ArgoDate;
 import fr.coriolis.checker.specs.ArgoReferenceTable;
